+++ src/main/java/org.team1/exceptions/ClientNotFoundException.java
@@ -2,7 +2,7 @@ package org.team1.exceptions;
 
 public class ClientNotFoundException extends RuntimeException {
 
-    public ClientNotFoundException(Integer amka) {
+    public ClientNotFoundException(Long amka) {
         super("Could not find client with amka " + amka);
     }
