+++ src/main/java/org.team1/exceptions/DoctorNotFoundException.java
@@ -2,7 +2,7 @@ package org.team1.exceptions;
 
 public class DoctorNotFoundException extends RuntimeException {
 
-    public DoctorNotFoundException(Integer amka) {
+    public DoctorNotFoundException(Long amka) {
         super("Could not find doctor with amka " + amka);
     }
 }
